Fix wrong certificate path in restund.conf
The task 'install restund tls certificate' puts the certificate
in a non-configurable path: /etc/restund/restund.pem. That path
is mounted 1:1 into the container. But since the Restund
configuration points to a non-existing PEM file path, the process
fails with during startup:

  Can't read certificate file /usr/local/etc/restunf/restund.pem

This change set adjusts the path in the Restund configuration file
to match the path used in the task creating the PEM file.